requirements.txt
asgiref==3.7.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==3.2.25
dnspython==2.3.0
eventlet==0.33.3
greenlet==3.0.1
idna==3.4
image==1.5.33
importlib-metadata==6.7.0
Nuitka==1.7.9
ordered-set==4.1.0
Pillow==9.5.0
prettytable==3.7.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.3.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.10
wifi-password==1.1.1
wincertstore==0.2
zipp==3.15.0
zstandard==0.21.0
